# Enhanced Data Availability Smart Contract