--- a/guideX_contracts/Clarinet.toml
+++ b/guideX_contracts/Clarinet.toml
@@ -9,6 +9,11 @@ requirements = []
 path = 'contracts/daily_update_limits.clar'
 clarity_version = 2
 epoch = 2.5
+
+[contracts.data_availability]
+path = 'contracts/data_availability.clar'
+clarity_version = 2
+epoch = 2.5
 [repl.analysis]
 passes = ['check_checker']