… guideX_contracts/contracts/data_availability.clar