<html>
<body>
<!--StartFragment--><html><head></head><body><h1>Enhanced Data Availability Smart Contract</h1>
<p>A comprehensive Clarity smart contract that provides enterprise-grade decentralized storage with advanced features including storage auctions, insurance, governance, CDN capabilities, and dynamic pricing mechanisms.</p>
<h2>Overview</h2>
<p>The Enhanced Data Availability contract creates a sophisticated decentralized storage network that goes beyond basic file storage to provide:</p>
<ul>
<li><strong>Decentralized Storage Network</strong>: Stake-based node participation with reputation systems</li>
<li><strong>Storage Auctions</strong>: Market-driven storage allocation</li>
<li><strong>Insurance System</strong>: File loss protection and claims processing</li>
<li><strong>Governance</strong>: Community-driven parameter management</li>
<li><strong>CDN Features</strong>: Regional caching and content delivery optimization</li>
<li><strong>Dynamic Pricing</strong>: Market-responsive storage costs</li>
<li><strong>Performance Metrics</strong>: Comprehensive node and network monitoring</li>
</ul>
<h2>Key Features</h2>
<h3>🏗️ Decentralized Storage Network</h3>
<ul>
<li><strong>Node Registration</strong>: Stake-based participation with geographic and specialization data</li>
<li><strong>Redundancy Management</strong>: Configurable file replication across multiple nodes</li>
<li><strong>Proof-of-Storage</strong>: Regular cryptographic proofs to verify file availability</li>
<li><strong>Reputation System</strong>: Performance-based node scoring and incentives</li>
</ul>
<h3>💰 Storage Marketplace</h3>
<ul>
<li><strong>Auction System</strong>: Competitive bidding for storage contracts</li>
<li><strong>Dynamic Pricing</strong>: Network utilization and regional demand-based pricing</li>
<li><strong>Retrieval Marketplace</strong>: Reward-based file retrieval requests</li>
<li><strong>Performance Incentives</strong>: Latency and bandwidth-based rewards</li>
</ul>
<h3>🛡️ Insurance &amp; Protection</h3>
<ul>
<li><strong>File Insurance</strong>: Optional coverage against file loss</li>
<li><strong>Claims Processing</strong>: Structured insurance claim system</li>
<li><strong>Slashing Mechanism</strong>: Stake penalties for node misbehavior</li>
<li><strong>Premium Pool</strong>: Community insurance fund management</li>
</ul>
<h3>🏛️ Governance System</h3>
<ul>
<li><strong>Proposal Creation</strong>: Stake-weighted governance proposals</li>
<li><strong>Community Voting</strong>: Democratic parameter and policy changes</li>
<li><strong>Parameter Updates</strong>: Dynamic system configuration</li>
<li><strong>Treasury Management</strong>: Community fund allocation</li>
</ul>
<h3>🌐 Content Delivery Network</h3>
<ul>
<li><strong>Regional Caching</strong>: Geographic content distribution</li>
<li><strong>Popular File Optimization</strong>: Automated caching of frequently accessed files</li>
<li><strong>Latency Optimization</strong>: Region-aware content delivery</li>
<li><strong>Cache Incentives</strong>: Rewards for CDN participation</li>
</ul>
<h2>Network Architecture</h2>
<h3>Node Types</h3>
<ul>
<li><strong>Full Nodes</strong>: Complete storage and retrieval capabilities</li>
<li><strong>Light Nodes</strong>: Limited storage with fast retrieval</li>
<li><strong>Archive Nodes</strong>: Long-term storage specialists</li>
<li><strong>Cache Nodes</strong>: CDN and regional distribution focus</li>
</ul>
<h3>Storage Tiers</h3>
<ul>
<li><strong>Hot Storage</strong>: High-frequency access, premium pricing</li>
<li><strong>Warm Storage</strong>: Medium access, balanced pricing</li>
<li><strong>Cold Storage</strong>: Archive storage, reduced pricing</li>
</ul>
<h3>Geographic Regions</h3>
<ul>
<li>Configurable regional deployment</li>
<li>Region-specific pricing multipliers</li>
<li>Latency optimization</li>
<li>Regulatory compliance support</li>
</ul>
<h2>Core Functions</h2>
<h3>Node Management</h3>
<h4><code>register-node-advanced</code></h4>
<pre><code class="language-clarity">(register-node-advanced 
  (stake-amount uint) 
  (region (string-ascii 10)) 
  (node-type (string-ascii 20)) 
  (specializations (list 5 (string-ascii 20))))
</code></pre>
<p>Register as a storage node with advanced configuration.</p>
<p><strong>Parameters:</strong></p>
<ul>
<li><code>stake-amount</code>: STX to stake (minimum 1 STX)</li>
<li><code>region</code>: Geographic region code</li>
<li><code>node-type</code>: "full", "light", "archive", or "cache"</li>
<li><code>specializations</code>: File types/services supported</li>
</ul>
<p><strong>Example:</strong></p>
<pre><code class="language-clarity">;; Register as a full node in US-East specializing in media files
(contract-call? .enhanced-data-availability register-node-advanced 
  u5000000      ;; 5 STX stake
  "us-east"     ;; Region
  "full"        ;; Full node
  (list "video" "audio" "images" "documents"))
</code></pre>
<h4><code>submit-proof-with-metrics</code></h4>
<pre><code class="language-clarity">(submit-proof-with-metrics 
  (file-hash (buff 32)) 
  (proof-hash (buff 32))
  (bandwidth-used uint)
  (latency-ms uint))
</code></pre>
<p>Submit proof-of-storage with performance metrics.</p>
<h3>File Management</h3>
<h4><code>register-file-advanced</code></h4>
<pre><code class="language-clarity">(register-file-advanced 
  (file-hash (buff 32)) 
  (content-hash (buff 32))
  (file-size uint) 
  (file-type (string-ascii 20))
  (encryption-type (string-ascii 20))
  (redundancy uint) 
  (priority uint)
  (expiry-blocks (optional uint))
  (access-permissions (list 10 principal))
  (insurance-coverage uint))
</code></pre>
<p>Register a file with comprehensive metadata and options.</p>
<p><strong>Parameters:</strong></p>
<ul>
<li><code>file-hash</code>: Unique file identifier</li>
<li><code>content-hash</code>: Integrity verification hash</li>
<li><code>file-size</code>: Size in bytes (max 1GB)</li>
<li><code>file-type</code>: Content type classification</li>
<li><code>encryption-type</code>: Encryption method used</li>
<li><code>redundancy</code>: Number of copies (minimum 3)</li>
<li><code>priority</code>: Priority level 1-5 (affects pricing)</li>
<li><code>expiry-blocks</code>: Optional automatic expiration</li>
<li><code>access-permissions</code>: Authorized access list</li>
<li><code>insurance-coverage</code>: Optional insurance value</li>
</ul>
<p><strong>Example:</strong></p>
<pre><code class="language-clarity">;; Register a critical document with insurance
(contract-call? .enhanced-data-availability register-file-advanced 
  0x1234567890abcdef1234567890abcdef12345678  ;; file-hash
  0xabcdef1234567890abcdef1234567890abcdef12  ;; content-hash
  u2500000    ;; 2.5MB file
  "document"  ;; File type
  "aes256"    ;; Encryption
  u5          ;; 5 copies
  u4          ;; High priority
  (some u52560)  ;; Expires in ~1 year
  (list 'ST1PQHQKV0RJXZFY1DGX8MNSNYVE3VGZJSRTPGZGM)  ;; Access list
  u1000000)   ;; 1 STX insurance coverage
</code></pre>
<h3>Storage Auctions</h3>
<h4><code>create-storage-auction</code></h4>
<pre><code class="language-clarity">(create-storage-auction (file-hash (buff 32)) (min-price uint) (required-specs (string-ascii 100)))
</code></pre>
<p>Create a competitive auction for file storage.</p>
<h4><code>place-auction-bid</code></h4>
<pre><code class="language-clarity">(place-auction-bid (target-auction-id uint) (bid-amount uint) (node-specs (string-ascii 100)))
</code></pre>
<p>Bid on a storage auction with node specifications.</p>
<p><strong>Example Auction Workflow:</strong></p>
<pre><code class="language-clarity">;; File owner creates auction
(contract-call? .enhanced-data-availability create-storage-auction 
  0x1234567890abcdef1234567890abcdef12345678
  u500000    ;; Minimum 0.5 STX
  "ssd,99.9%,us-west")  ;; SSD storage, 99.9% uptime, US-West region

;; Nodes bid on the auction
(contract-call? .enhanced-data-availability place-auction-bid 
  u1         ;; Auction ID
  u600000    ;; 0.6 STX bid
  "nvme,99.95%,10ms")  ;; NVMe, 99.95% uptime, 10ms latency
</code></pre>
<h3>Retrieval Marketplace</h3>
<h4><code>request-file-retrieval</code></h4>
<pre><code class="language-clarity">(request-file-retrieval 
  (file-hash (buff 32)) 
  (reward-offered uint) 
  (max-latency uint) 
  (preferred-region (optional (string-ascii 10))))
</code></pre>
<p>Request fast file retrieval with performance requirements.</p>
<h4><code>fulfill-retrieval-request</code></h4>
<pre><code class="language-clarity">(fulfill-retrieval-request (target-request-id uint) (delivery-time uint))
</code></pre>
<p>Fulfill a retrieval request and claim the reward.</p>
<p><strong>Example:</strong></p>
<pre><code class="language-clarity">;; Request urgent file retrieval
(contract-call? .enhanced-data-availability request-file-retrieval 
  0x1234567890abcdef1234567890abcdef12345678
  u100000     ;; 0.1 STX reward
  u5000       ;; Max 5 second latency
  (some "us-east"))  ;; Prefer US-East region

;; Node fulfills the request
(contract-call? .enhanced-data-availability fulfill-retrieval-request 
  u1          ;; Request ID
  u3500)      ;; Delivered in 3.5 seconds
</code></pre>
<h3>Insurance System</h3>
<h4><code>file-insurance-claim</code></h4>
<pre><code class="language-clarity">(file-insurance-claim 
  (file-hash (buff 32)) 
  (claim-amount uint) 
  (reason (string-ascii 50)) 
  (evidence-hash (buff 32)))
</code></pre>
<p>File an insurance claim for lost or corrupted files.</p>
<p><strong>Example:</strong></p>
<pre><code class="language-clarity">;; File insurance claim for data loss
(contract-call? .enhanced-data-availability file-insurance-claim 
  0x1234567890abcdef1234567890abcdef12345678
  u800000     ;; Claim 0.8 STX
  "node-failure-data-loss"
  0xdeadbeefcafebabe1234567890abcdef12345678)  ;; Evidence hash
</code></pre>
<h3>Governance</h3>
<h4><code>create-governance-proposal</code></h4>
<pre><code class="language-clarity">(create-governance-proposal 
  (title (string-ascii 100)) 
  (description (string-ascii 500))
  (proposal-type (string-ascii 20))
  (target-parameter (optional (string-ascii 50)))
  (new-value (optional uint)))
</code></pre>
<p>Create governance proposals for network parameters.</p>
<h4><code>vote-on-proposal</code></h4>
<pre><code class="language-clarity">(vote-on-proposal (proposal-id uint) (vote bool))
</code></pre>
<p>Vote on governance proposals (stake-weighted).</p>
<p><strong>Example:</strong></p>
<pre><code class="language-clarity">;; Propose increasing minimum stake
(contract-call? .enhanced-data-availability create-governance-proposal 
  "Increase Minimum Stake"
  "Proposal to increase minimum node stake from 1 STX to 2 STX to improve network security and reduce spam nodes"
  "parameter"
  (some "min-stake")
  (some u2000000))  ;; 2 STX

;; Vote on the proposal
(contract-call? .enhanced-data-availability vote-on-proposal u1 true)  ;; Vote YES
</code></pre>
<h3>CDN Caching</h3>
<h4><code>cache-popular-file</code></h4>
<pre><code class="language-clarity">(cache-popular-file (file-hash (buff 32)) (region (string-ascii 10)))
</code></pre>
<p>Cache frequently accessed files in regional nodes.</p>
<h2>Data Structures</h2>
<h3>Node Information</h3>
<pre><code class="language-clarity">{
  stake: uint,
  reputation: uint,
  last-proof-block: uint,
  is-active: bool,
  slash-count: uint,
  total-files: uint,
  bandwidth-used: uint,
  geographic-region: (string-ascii 10),
  node-type: (string-ascii 20),
  performance-score: uint,
  uptime-percentage: uint,
  last-ping: uint,
  specializations: (list 5 (string-ascii 20))
}
</code></pre>
<h3>File Metadata</h3>
<pre><code class="language-clarity">{
  owner: principal,
  size: uint,
  created-block: uint,
  required-redundancy: uint,
  current-redundancy: uint,
  reward-per-copy: uint,
  file-type: (string-ascii 20),
  encryption-type: (string-ascii 20),
  access-frequency: uint,
  priority-level: uint,
  insurance-value: uint,
  retrieval-count: uint,
  content-hash: (buff 32),
  expiry-block: (optional uint),
  access-permissions: (list 10 principal)
}
</code></pre>
<h3>Storage Auction</h3>
<pre><code class="language-clarity">{
  file-hash: (buff 32),
  creator: principal,
  min-price: uint,
  current-winner: (optional principal),
  current-bid: uint,
  end-block: uint,
  is-active: bool,
  required-specs: (string-ascii 100)
}
</code></pre>
<h2>Read-Only Functions</h2>
<h3>Network Information</h3>
<ul>
<li><code>get-network-stats()</code> - Overall network statistics</li>
<li><code>get-regional-metrics(region)</code> - Region-specific metrics</li>
<li><code>calculate-slash-amount(node-id)</code> - Potential slash amount for node</li>
</ul>
<h3>Node &amp; File Details</h3>
<ul>
<li><code>get-node-detailed-info(node-id)</code> - Complete node information</li>
<li><code>get-file-detailed-info(file-hash)</code> - Complete file metadata</li>
<li><code>get-auction-info(auction-id)</code> - Auction details</li>
</ul>
<h3>Insurance &amp; Governance</h3>
<ul>
<li><code>get-insurance-policy(file-hash)</code> - Insurance policy details</li>
<li><code>get-governance-proposal(proposal-id)</code> - Proposal information</li>
</ul>
<h2>Economic Model</h2>
<h3>Staking Economics</h3>
<ul>
<li><strong>Minimum Stake</strong>: 1 STX per node</li>
<li><strong>Slashing Rate</strong>: 20% of stake for misbehavior</li>
<li><strong>Reputation Impact</strong>: Performance affects future rewards</li>
</ul>
<h3>Storage Pricing</h3>
<ul>
<li><strong>Base Price</strong>: Dynamic based on network utilization</li>
<li><strong>Priority Multiplier</strong>: Higher priority = higher cost</li>
<li><strong>Regional Adjustment</strong>: Supply/demand based regional pricing</li>
<li><strong>Insurance Premium</strong>: 5% of coverage value</li>
</ul>
<h3>Reward Distribution</h3>
<ul>
<li><strong>Proof Submission</strong>: 0.001 STX per proof</li>
<li><strong>Redundancy Bonus</strong>: 0.1 STX per file copy</li>
<li><strong>Retrieval Rewards</strong>: Market-determined by requests</li>
<li><strong>CDN Caching</strong>: 0.05 STX for popular file caching</li>
</ul>
<h2>Security Features</h2>
<h3>Cryptographic Security</h3>
<ul>
<li><strong>Proof-of-Storage</strong>: Regular cryptographic proofs required</li>
<li><strong>Content Integrity</strong>: SHA-256 checksums for verification</li>
<li><strong>Access Control</strong>: Permission-based file access</li>
</ul>
<h3>Economic Security</h3>
<ul>
<li><strong>Stake Slashing</strong>: Financial penalties for misbehavior</li>
<li><strong>Reputation System</strong>: Long-term performance tracking</li>
<li><strong>Insurance Fund</strong>: Community-backed loss protection</li>
</ul>
<h3>Network Security</h3>
<ul>
<li><strong>Redundancy Requirements</strong>: Minimum 3 copies per file</li>
<li><strong>Geographic Distribution</strong>: Multi-region storage</li>
<li><strong>Performance Monitoring</strong>: Continuous latency and bandwidth tracking</li>
</ul>
<h2>Configuration Parameters</h2>
<h3>Network Limits</h3>
<ul>
<li><strong>Maximum File Size</strong>: 1GB</li>
<li><strong>Bandwidth Limit</strong>: 10MB per block per node</li>
<li><strong>Auction Duration</strong>: 7 days (1,008 blocks)</li>
<li><strong>Proof Interval</strong>: 24 hours (144 blocks)</li>
</ul>
<h3>Economic Parameters</h3>
<ul>
<li><strong>Minimum Stake</strong>: 1,000,000 microSTX (1 STX)</li>
<li><strong>Slash Percentage</strong>: 20%</li>
<li><strong>Insurance Premium Rate</strong>: 5%</li>
<li><strong>Governance Quorum</strong>: 51%</li>
</ul>
<h3>Performance Thresholds</h3>
<ul>
<li><strong>Reputation Decay</strong>: 1 point per missed proof</li>
<li><strong>Performance Bonus</strong>: +5 points for &lt;100ms latency</li>
<li><strong>Cache Popularity</strong>: 10+ accesses required</li>
</ul>
<h2>Error Codes</h2>

Code | Name | Description
-- | -- | --
100 | ERR_UNAUTHORIZED | Insufficient permissions
101 | ERR_NODE_NOT_FOUND | Node doesn't exist
102 | ERR_FILE_NOT_FOUND | File not registered
103 | ERR_INSUFFICIENT_STAKE | Stake below minimum
104 | ERR_PROOF_EXPIRED | Proof submission too late
105 | ERR_INVALID_PROOF | Cryptographic proof invalid
106 | ERR_ALREADY_SLASHED | Node already penalized
107 | ERR_INSUFFICIENT_BALANCE | Not enough STX
108 | ERR_CONTRACT_PAUSED | Emergency pause active
109 | ERR_INVALID_SIGNATURE | Signature verification failed
110 | ERR_AUCTION_ENDED | Auction period closed
111 | ERR_BID_TOO_LOW | Bid below minimum/current
112 | ERR_INSURANCE_CLAIM_DENIED | Claim not valid
113 | ERR_BANDWIDTH_EXCEEDED | Node bandwidth limit hit
114 | ERR_INVALID_ENCRYPTION | Encryption type unsupported
115 | ERR_GOVERNANCE_THRESHOLD | Insufficient stake for governance


<h2>Best Practices</h2>
<h3>For Node Operators</h3>
<ul>
<li>Maintain high uptime and fast response times</li>
<li>Submit proofs regularly to maintain reputation</li>
<li>Specialize in specific file types for better rewards</li>
<li>Monitor bandwidth usage to avoid limits</li>
<li>Participate in governance to influence network direction</li>
</ul>
<h3>For File Owners</h3>
<ul>
<li>Choose appropriate redundancy levels for importance</li>
<li>Use insurance for critical files</li>
<li>Set realistic expiry dates</li>
<li>Monitor access patterns for optimization</li>
<li>Participate in auctions for cost efficiency</li>
</ul>
<h3>For Developers</h3>
<ul>
<li>Implement proper error handling for all operations</li>
<li>Monitor network metrics for optimal performance</li>
<li>Use retrieval marketplace for urgent needs</li>
<li>Cache popular files regionally for better UX</li>
<li>Integrate governance notifications for parameter changes</li>
</ul>
<h2>Integration Examples</h2>
<h3>File Upload Service</h3>
<pre><code class="language-clarity">;; Complete file upload with insurance
(define-public (upload-critical-file (file-data (buff 32)) (file-size uint))
  (let (
    (content-hash (sha256 file-data))
    (insurance-amount (* file-size u1000)) ;; 1000 microSTX per byte
  )
    (contract-call? .enhanced-data-availability register-file-advanced 
      file-data
      content-hash
      file-size
      "document"
      "aes256"
      u5          ;; 5 copies for critical files
      u5          ;; Highest priority
      (some u52560)  ;; 1 year expiry
      (list tx-sender)  ;; Only owner access
      insurance-amount)
  )
)
</code></pre>
<h3>CDN Integration</h3>
<pre><code class="language-clarity">;; Auto-cache popular files
(define-public (auto-cache-popular-files (region (string-ascii 10)))
  ;; Implementation would query popular files and cache them
  (ok true)
)
</code></pre>
<h3>Monitoring Dashboard</h3>
<pre><code class="language-clarity">;; Get comprehensive network status
(define-read-only (get-dashboard-data)
  (let (
    (network-stats (contract-call? .enhanced-data-availability get-network-stats))
    (regional-data (contract-call? .enhanced-data-availability get-regional-metrics "us-east"))
  )
    {
      network: network-stats,
      regions: regional-data,
      current-block: block-height
    }
  )
)
</code></pre>
<h2>Deployment Guide</h2>
<ol>
<li><strong>Deploy Contract</strong>: Deploy to Stacks blockchain</li>
<li><strong>Initialize Parameters</strong>: Set network configuration</li>
<li><strong>Bootstrap Nodes</strong>: Register initial storage nodes</li>
<li><strong>Configure Regions</strong>: Set up geographic regions</li>
<li><strong>Fund Insurance Pool</strong>: Initial insurance fund</li>
<li><strong>Enable Governance</strong>: Activate community governance</li>
…